requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.7.4
stripe==10.11.0
qrcode[pil]==7.4.2
pillow==10.4.0
python-multipart==0.0.9
